last changes

+++ lib/Features/Task/data/models/task_model.dart
@@ -1,10 +1,10 @@
-import 'dart:ffi';
+
 import 'package:clean_arch_task_app/Features/Task/domain/entities/task_entitie.dart';
 
 class TaskModel extends TaskEntitie {
   int taskId;
   String taskName;
-  Short taskStatus;
+  int taskStatus;
 
   TaskModel(
       {required this.taskId, required this.taskName, required this.taskStatus})
