+++ lib/Features/Task/domain/entities/task_entitie.dart
@@ -1,9 +1,9 @@
-import 'dart:ffi';
+
 
 class TaskEntitie {
   int taskId;
   String taskName;
-  Short taskStatus;
+  int taskStatus;
 
   TaskEntitie({required this.taskId,required this.taskName, required this.taskStatus});
 }
